Add support for FIP SP packaging method
This commit introduces the SP_PACKAGING_METHOD config option, which
decides how the SP images built by TS are represented in the system.
Currently the two supported options are "embedded" and "fip". The former
means the method used so far (including the SP images in the main OP-TEE
binary). The latter means creating SP packages which are included as
separate entries in the FIP, and will be loaded by BL2 at boot. This
method relies on DT files to inform BL2 and SPMC about the available SP
packages. The commit also contains an example for both of these files.

Warning: the ARM_BL2_SP_LIST_DTS config option for TF-A is used when
selecting the FIP packaging method. However, currently this is not
available in upstream TF-A, only on a development branch.

Change-Id: Ic3a5dc906241b7dc944d4eca66f4f389a7c935cb
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>