Fix FF-A SP initialization order
When using the "embedded" booth method initialization order is
specified by the order SPs are placed on the linker command line.
The original implementation used parallel dependency between targets
including the makefiles exported from Trusted Services. As a result
the initialization order was decided by make, and was not possible to
be changed.
This patch uses gnumake's "makefile remake" feature. The solution
sets the link order (and thus the initialization order) to match the
order SPs are mentioned in the makefile.

Note: this is a workaround for the limitation of the SPMC no being
      able to handle the "Boot order" filed in the FF-A manifest.

Change-Id: I18158b7cab17302bbfd1764e0140fdd609f93f60
Signed-off-by: Gyorgy Szing Gyorgy.Szing@arm.com